[IMP] udes_stock: Tidy up orderpoint model
Merge multiple model extensions into one file.
