+++ addons/udes_stock/models/stock_warehouse_orderpoint.py
@@ -1,4 +1,5 @@
-from odoo import fields, models, _, api
+from odoo import fields, models, _, api, registry
+from odoo.osv import expression
 from odoo.exceptions import ValidationError
 
 
@@ -67,3 +68,34 @@ class Orderpoint(models.Model):
         Product._compute_nbr_reordering_rules()
         return res
 
+    @api.model
+    def check_order_points(self, use_new_cursor=False, company_id=False, location_id=False, excluded_location_ids=False):
+        """
+        Copy of run_scheduler from Odoo's stock module ProcurementGroup class.
+        This allows us to only check order points.
+        """
+        OrderPoint = self.env["stock.warehouse.orderpoint"]
+        ProcurementGroup = self.env["procurement.group"]
+        try:
+            if use_new_cursor:
+                domain = ProcurementGroup._get_orderpoint_domain()
+                if location_id:
+                    domain = expression.AND([domain, [("location_id", "=", location_id)]])
+                if excluded_location_ids:
+                    domain = expression.AND([domain, [("location_id", "not in", excluded_location_ids)]])
+                self = OrderPoint.with_context(prefetch_fields=False).search(domain)
+                cr = registry(self._cr.dbname).cursor()
+                self = self.with_env(self.env(cr=cr))
+
+            self.sudo()._procure_orderpoint_confirm(
+                use_new_cursor=use_new_cursor,
+                company_id=company_id)
+            if use_new_cursor:
+                self._cr.commit()
+        finally:
+            if use_new_cursor:
+                try:
+                    self._cr.close()
+                except Exception:
+                    pass
+        return {}
